+++ platform/viewer/src/connectedComponents/ConnectedStudyBrowser.js
@@ -20,12 +20,20 @@ const mapDispatchToProps = (dispatch, ownProps) => {
         const { Modality } = displaySet;
         if (Modality === 'SEG' && servicesManager) {
           const onDisplaySetLoadFailureHandler = error => {
-            LoggerService.error({ error, message: error.message });
+            const message =
+              error.message.includes('orthogonal') ||
+              error.message.includes('oblique')
+                ? 'The segmentation has been detected as non coplanar,\
+                If you really think it is coplanar,\
+                please adjust the tolerance in the segmentation panel settings (at your own peril!)'
+                : error.message;
+
+            LoggerService.error({ error, message });
             UINotificationService.show({
               title: 'DICOM Segmentation Loader',
-              message: error.message,
+              message,
               type: 'error',
-              autoClose: true,
+              autoClose: false,
             });
           };
 
@@ -46,7 +54,9 @@ const mapDispatchToProps = (dispatch, ownProps) => {
                 detail: { activatedLabelmapIndex: activatedLabelmapIndex },
               }
             );
+            const segThumbnailSelected = new CustomEvent('segseriesselected');
             document.dispatchEvent(selectionFired);
+            document.dispatchEvent(segThumbnailSelected);
           });
         } else if (Modality !== 'SR') {
           displaySet = displaySet.getSourceDisplaySet(ownProps.studyMetadata);